extract logout to service object
