+++ .env.sample
@@ -77,4 +77,3 @@ ANTS_RDV_API_URL="https://int.api-coordination.rendezvouspasseport.ants.gouv.fr/
 AGENT_CONNECT_BASE_URL="https://fca.integ01.dev-agentconnect.fr/api/v2"
 AGENT_CONNECT_CLIENT_ID="ec41582-1d60-4f11-a63b-d8abaece16aa"
 AGENT_CONNECT_CLIENT_SECRET="voir https://coffre.incubateur.anct.gouv.fr/#/vault?folderId=577313a3-40e2-48e9-b043-273019afc22f&itemId=48247d0e-fc14-47b4-b3bb-11c78f77d1a7"
-
